Correct unit and system tests
Moved bootstrap tests to unpackaged.
Corrected runScript() on system tests so that
they no longer use bootstrap inputs.

Moved all unit and system tests for analysis and calibration into its own folder.
Changed paths of imports to account for new structure

Had to change github workflows to include installing the
mvesuvio package so that imports from unit and system
tests work.

Added a default package version to pyproject.toml to fall back to if git is not found.
This is necessary for when the mvesuvio package is built during
the github workflows so it doesn't stop the workflow with an error
message.