Run coverage nightly and send to coveralls
The nightly unit tests are now set to run with coverage.
The coverage report is sent to coveralls.
The host was changed from self-hosted to ubuntu-latest because
it decreases the time of queue for jobs.
The configuration of coverage sets coverage to run only on
the specified files and ignores other files.
Added badge to readme.md file to show test coverage.

For coverage to find the correct files, need to use
an editable pip install by using the -e flag, otherwise
coverage will not find any files that are packaged inside of src/