Put set up of fit parameters in class methods
Initialize fit parameters and bounds when setting up the class.
This makes it easier to keep track of the order in which the masses are
getting into the parameters array and makes it easier to change it in the
future.

Made self._masses method because it is used several times across the routine.