Implement passing constraints as arg
Previously the constraints argument was disabled because
I did not find a way to pass Python objects (tuple of dictionaries) into
the mantid algorithm.

This workaround converts the tuple of dictionaries into a byte string
using the dill library which can then be converted to a normal string and
passed into the mantid algorithm as an argument.
When the algorithm executes, the string is converted into byte string again
though eval() and then into the tuple of dicts of lambdas by the dill library.

A better solution might be implemented in the future but this will suffice for now.