Extend y space range for smoother derivative
The function for the numerical third derivative trims the range of x by
6 columns on each side. In order to match the shape of the data, the
solution is to extend the range of y space so that both y space and the
resulting JOfY function have 6 more points than needed on each side.

After the third derivative is computed, trim JOfY to match shape of FSE.

This solution avoids padding the FSE corrections with zeros, and
maintains the same computational speed.