Add test for calculating means
Cleaned function for calculating means in the reduction analysis
and added unit test to check the filtering of widths and intensities is working.

I removed all of the assertions because instead of stoping the code it's best to
at least keep going and produce the table for means so that at least the user
can have a clue for what happened.