Run mvesuvio before uploading package
During the nightly deployment workflow, it's best to test
the basic functunality of config and make sure that
at least this command does not give an error.

In the case of conda, this is best done in the conda recipe,
by adding a test section that runs the command for config.
(Also added command to run unit tests)