Improve third derivative (#148)
Previous third derivative result was padded with 6 zeros on each side.
This improvement extends the range of the kinematic, y-space and resolution arrays by 6 indices on each side, so that when the third derivative is taken, the size of the resulting array is 12 indices less than the input (6 on each side).

So in essence this just means that I am extending the range before taking the derivative so that after the derivative the size matches the size before the extension. This makes sure the derivative is continuous and does not including zero padding, which contributes to a smoother ncp profile.

Added and updated unit tests. Updated benchmark for system tests.