Change name of log file and print tables
Changed the name of the log file to account for the fact that one
of the routines is joint, so hte name cannot depend on the spectra being
fitted (well it could, but that would be unecessary and more complex).

I also noticed that no information about the initial parameters was being
outputed so I created a new function for printing tables and used that to
show the tables for the initial parameters.