Improve printing of tables
Convert to integers where possible so that priting of zeros is avoided.
Changed names of columns of fit parameters because they were taking too
much space when the table got print, which made it very difficult to read.
Now table occupies much less space and is much clearer to read.