Add check for clearing log only on linux
The log file cannot be cleaned when on Windows, because due
to file locks, cannot open file when it is being used by Mantid
to write to it.

The best solution for now is to clean the file when on Unix system.
On windows the log file will still be produced, but will contain the logs
of all times the routines were run.