Replace system test for reduction
Previous system tests were using a file saved in npz, which is not used by the users and was quite cumbersome to work with.
This change outputs the relevant workspaces to a reduction folder (useful for users + useful for testing).
It also means that the code for testing the workspaces becomes much simpler and testing becomes more comprehensive.