Create log file vesuvio (#153)
* Add logging settings to mantid properties file

Logging settings are added during config command.
Current downside is that if logging settings are set to
mantid properties when mantid is open, mantid needs to be
closed and opened again for changes to take place.

* Summarize mantid log into a new file

Added function that takes in existing mantid log file and
produces a clean version with only the important information.

* Add output from fitting to log file

Replaced print statements in fitting with logger.notice().
Added these statements to log file.

* Change name of log file and print tables

Changed the name of the log file to account for the fact that one
of the routines is joint, so hte name cannot depend on the spectra being
fitted (well it could, but that would be unecessary and more complex).

I also noticed that no information about the initial parameters was being
outputed so I created a new function for printing tables and used that to
show the tables for the initial parameters.

* Improve printing of tables

Convert to integers where possible so that priting of zeros is avoided.
Changed names of columns of fit parameters because they were taking too
much space when the table got print, which made it very difficult to read.
Now table occupies much less space and is much clearer to read.

* Replace old notices with print tables

Previous notices were not very clear and made the code more
confusing, replaced them with printing tables that are already
bring created anyway.
Log output is much easier to follow this way

* Add unit test for creating log file

Unit test for making a summarised log file from mantid log file

* Add check for clearing log only on linux

The log file cannot be cleaned when on Windows, because due
to file locks, cannot open file when it is being used by Mantid
to write to it.

The best solution for now is to clean the file when on Unix system.
On windows the log file will still be produced, but will contain the logs
of all times the routines were run.