Fix masked TOF columns in reduction (#173)
* Correct mistake when masking values in error function

Previous error function was calculating the nonzero values in an incorrect manner.
this is because I was modifying data_y array and then using it again with np.nonzero().
This was clearly a blunder on my part and was producing bad fits when some values are masked with zeros.
The change is now covered by the unit test