Refactor System tests (#169)
* Split yspace class into back and front

* Take out part for isolating lightest mass and resolution

Took out function for isolating lightest mass data out of the fitting
algorithm. Also took out resolution function.
This makes more sense as the algorithm should act on the workspace that
is going to fit.

* Changed implementation of isolating the lightest mass

Previous implementation was using Minus(), which I realised changes the
Error values due to error propagation.
So I changed the implementation to only change the dataY values

* Add unit tests for isolating mass and resolution

Since I took the functions for isolating the mass and calculating the
resolution out of the y-space fit, I added these unit tests since I am
not changing their implementation from now onwards.

* Replace system test for reduction

Previous system tests were using a file saved in npz, which is not used by the users and was quite cumbersome to work with.
This change outputs the relevant workspaces to a reduction folder (useful for users + useful for testing).
It also means that the code for testing the workspaces becomes much simpler and testing becomes more comprehensive.

* Set up system test for fitting routine

The results of the benchmark need to be updated because currently the
fit in y space is not dealing well with zeros in TOF

* Correct function for isolating mass to account for zeros

In my previous implementation of the function to isolate the mass I forgot to account for masked values with zeros.

* Remove last column of fit results table

The comparison between fit results is failing on github runners, and I have a suspicion that the number of iterations might differ.
This column is not really relevant anyway, so I removed it to see if it passes the tests on github.