Add plot config (#199)
* Add arg to override input workspace

When running analysis, the input workspace can now be overriden by
passing an extra argument containing the path of the file to override.
This means that empty or raw runs will not be loaded and workspace is
assumed to have been cropped already.

* Limit overriding workspaces to nexus files

If the workspace is loaded using LoadAscii it will not contain geometry
information necessary or multiscattering and gamma corrections

* Add arg for enabling minimal output

Will need this option in bootstrap, where hundreds of replicas will be
created

* Add bootstrap routine

Iterates over inputs directory and generates results in output directory

* Add y space fitting to bootstrap

- As of now bootstrap runs the entire routine, i.e. analysis + fitting
- Cleaned up run() method in analysis runner to be more versatile and
easier to understand
- Cleaned up search for workspaces that are meant to be fitted

* Disable minos and global fit when errors are zero

If the workspace has zero errors, do not run minos or global fit since
it's currently failing

* Clean setting up of config folder

The default config directory should not be exposed to the user (at least
for now), which simplifies the setup of the config dir

* Add backward and forward options for bootstrap

Do the same routine as in joint but either only for backward or forward.
Save paths using .absolute(), otherwise they're not saved in the right
location